Add files via upload
This simple restaurant website is developed through the use of HTML, CSS as well as a pinch of JavaScript. To begin with, it consists of a welcoming home page with captivating images and a short description about the hotel. Additionally, it includes another page containing a comprehensive list of food items including their categories and prices. Finally, there is also a contact page that provides information about where the restaurant is situated; its telephone number is also given here – even though there isn’t any other relevant text or details about this business establishment with just an empty contact form for questions. The purpose of this example was to demonstrate basic web design knowledge and an understanding of how websites work.